+++ test/mithril.redraw.js
@@ -31,8 +31,6 @@ describe("m.redraw()", function () {
 	})
 
 	it("runs unnecessary redraws asynchronously", function () {
-		if (mock.phantom) return
-
 		var root = mock.document.createElement("div")
 		var view = sinon.spy()
 
@@ -198,8 +196,6 @@ describe("m.redraw()", function () {
 		})
 
 		dit("recreates the component when \"all\"", function (root) {
-			if (mock.phantom) return
-
 			var count = 0
 			function config(el, init) {
 				if (!init) count++